requirements.txt
google-genai
assemblyai
python-dotenv
pydub
google-api-python-client
google-auth-httplib2
google-auth-oauthlib